Add mounting rail on TOP of cantilever mount
I don't love this but I'm trying to prevent the screws' turning from
distorting the keys' gutters. This was also a problem w/ the OKAYs.

Anyway, this kinda seems to work but isn't 100%.

Also has some magic in it to reduce the height arbitrarily, because the
ideal math was just a litttle too tall and prevent lid from closing
all the way.