Re-annotate to fix key vs 555 mismatch
The alignment of switch to 555 circuit was off, starting around S8. I
believe this was from non-linear copy-pasting when key count jumped from
13 to 20.

No amound of fiddling with the annotation tool seemed to fix and to do
it manually was out of the question, so I ended up deleting 6-20 and
copy-pasting 1-5 three times. I think it worked...