Fix poor hitch positioning
Hitch was 1) too close to the end of the accidental for it to properly
enclose on both sides and 2) too low into the key for it to catch.

This change 1) lengthens accidentals based on measurements from my
Hammond and 2) exposes + uses a HITCH_RECOMMENDED_MINIMUM_CAVITY_HEIGHT
const instead of some arbitrary, key-derived thing.