Cheaply try to strengthen enclosure
Because the enclosure's inner cavity isn't rounded, there was a slight
gap in the lip's other wall where its rounded edge was cut off, and it
was WAY too easy for it to chip off after printed.

A proper fix would be to round the cavity and maybe increase the wall
depth or add some other reinforcement.

* Decreases lip (which is on the inside) so the outer part can be wider
* Switches lip from bottom to top -- w/o the window in its cavity to
  hold the width, the top can be squeezed in because there's no lip
  there
* Adjusts bottom height to include lip -- now that the bottom has a
  bumper that prevents easily sliding the keys onto their hitch, this
  height is arbitrary, but whatever